requirements.txt
flask
flask-socketio
playwright
pytest-playwright
requests
supabase
python-dotenv
eventlet
python-telegram-bot[job-queue]
gunicorn
